Add files via upload
This Python script uses Scikit Image for the image processing, Scikit Learn Gaussian Mixture Model (GMM) to analyse the distribution of gray levels, and Matplotlib and cv2 to plot the outputs. Basically, it shows that using the GMM model is pretty useful to detect the gary levels modes, it prevents from a manual selection of the threshold levels that would be user-dependent, like it is the case in ImageJ. Hopefully, it should work not too bad for more that two phases, provided that the gray levels would be sufficiently distinct.
Hope you'll enjoy! Don't hesitate to suggest some possible improvements or to contribute to it.